requirements.txt
aiogram
g4f
python-pptx